+++ chartlets.py/chartlets/components/__init__.py
@@ -1,6 +1,6 @@
 from .box import Box
 from .button import Button
 from .checkbox import Checkbox
-from .dropdown import Dropdown
 from .plot import Plot
+from .select import Select
 from .typography import Typography